Add create release example
The current "Deploy on publishing a new release and attach a ZIP file to the release" example uses `actions/upload-release-asset` which is deprecated.

Attaching files to a previous release doesn't seem that common in my research, the maintained GitHub release actions seem to focus on creating a release, optionally with files e.g. `softprops/action-gh-release`.

This adds a "Deploy on pushing a new tag and create release with attached ZIP" example to the README which seems to be the latest approach.

The implication with creating the release in an action is the changelog will be generated in the action too. I've omitted that from my example but would be happy to include if you like, perhaps there could be an `/examples` folder in this repo.