Update deploy.sh to fix shell glob in svn propsets
The glob of `*.png`, `*.jpg`, `*.gif`, and `*.svg` in the `svn propset svn:mime-type` commands fail due to being quoted. SVN does not support expanding the globs itself, so we need to rely on the shell expanding them. Shell globs are only expanded when the `*` is unquoted, but currently the whole path including the glob is within quotes meaning that we're trying to pass the literal string `*.png` etc with the `*` unexpanded.

* Unquote the globs in `svn propset svn:mime-type` commands to allow expansion by the shell